+++ tests/cases/lib/types/page-test.php
@@ -290,6 +290,9 @@ class Papi_Lib_Types_Page_Test extends WP_UnitTestCase {
 		$this->assertTrue( ! empty( $actual ) );
 		$this->assertTrue( is_array( $actual ) );
 
+		$slugs = papi_get_slugs( $this->post_id, true );
+		$this->assertTrue( array_filter( $slugs, 'is_string' ) === $slugs );
+
 		update_post_meta( $this->post_id, papi_get_page_type_key(), '' );
 		$this->flush_cache();
 		$this->assertEmpty( papi_get_slugs() );